feat: add modular structure
