+++ README.md
@@ -16,8 +16,3 @@
 **Figma**
 
 - [Ссылка на макет в Figma](https://www.figma.com/file/2cn9N9jSkmxD84oJik7xL7/JavaScript.-Sprint-4?node-id=0%3A1)
-
-**Планы по улучшению**
-
-1. Добавить transition всем интерактивным элементам с ховером для плавности анимации наведения.
-2. Всем кнопкам без текстового контента добавить атрибут aria-label.
