+++ src/blocks/content/content.css
@@ -10,6 +10,7 @@
 @media screen and (max-width: 992px) {
   .content {
     max-width: 728px;
+    align-items: center;
   }
 }
