+++ src/blocks/popup/__container/_modal/popup__container_modal_form.css
@@ -3,3 +3,9 @@
   width: 100%;
   z-index: 2;
 }
+
+@media screen and (max-width: 550px) {
+  .popup__container_modal_form {
+    max-width: 282px;
+  }
+}
