+++ src/blocks/profile/__add-new-button/profile__add-new-button.css
@@ -13,6 +13,7 @@
 @media screen and (max-width: 992px) {
   .profile__add-new-button {
     width: 100%;
+    min-width: 282px;
   }
 }
