+++ src/components/utils.js
@@ -1,0 +1,4 @@
+/* -------- Toggling Like Button -------- */
+export function toggleLikeButton(evt) {
+  evt.target.classList.toggle("element__like-button_active");
+}
