+++ src/components/validate.js
@@ -1,10 +1,68 @@
-// Вынесем все необходимые элементы формы в константы
-const formElement = document.querySelector('.popup__form_profile');
-const formInput = formElement.querySelector('.popup__input');
+/* -------- Display and Hide Input Errors -------- */
+function showInputError(formElement, inputElement, errorMessage, config) {
+  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+  inputElement.classList.add(config.inputErrorClass);
+  errorElement.classList.add(config.errorClass);
+  errorElement.textContent = errorMessage;
+};
 
-// Слушатель события input
-formInput.addEventListener('input', function (evt) {
-  // Выведем в консоль значение свойства validity.valid поля ввода,
-  // на котором слушаем событие input
-  console.log(evt.target.validity.valid);
-});
+function hideInputError(formElement, inputElement, config) {
+  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+  inputElement.classList.remove(config.inputErrorClass);
+  errorElement.classList.remove(config.errorClass);
+  errorElement.textContent = '';
+};
+
+/* -------- Check Inputs Validity & Show/Hide Error Message -------- */
+function isValid(formElement, inputElement, config) {
+  if (inputElement.validity.patternMismatch) {
+    inputElement.setCustomValidity(inputElement.dataset.errorMessage);
+  } else {
+    inputElement.setCustomValidity('');
+  }
+
+  if (!inputElement.validity.valid) {
+    showInputError(formElement, inputElement, inputElement.validationMessage, config);
+  } else {
+    // Если проходит, скроем
+    hideInputError(formElement, inputElement, config);
+  }
+};
+
+/* -------- Check Altogether Validity to Unblock Submit Button -------- */
+function hasInvalidInput(inputList) {
+  return inputList.some((inputElement) => {
+    return !inputElement.validity.valid;
+  })
+};
+
+function toggleButtonState(inputList, buttonElement, config) {
+  if (hasInvalidInput(inputList)) {
+    buttonElement.disabled = true;
+    buttonElement.classList.add(config.inactiveButtonClass);
+  } else {
+    buttonElement.disabled = false;
+    buttonElement.classList.remove(config.inactiveButtonClass);
+  }
+};
+
+/* -------- Set Validation Event Listeners -------- */
+function setEventListeners(formElement, config) {
+  const inputList = Array.from(formElement.querySelectorAll(config.inputSelector));
+  const buttonElement = formElement.querySelector(config.submitButtonSelector);
+  toggleButtonState(inputList, buttonElement, config);
+  inputList.forEach((inputElement) => {
+    inputElement.addEventListener('input', () => {
+      isValid(formElement, inputElement, config);
+      toggleButtonState(inputList, buttonElement, config);
+    });
+  });
+};
+
+/* -------- Enable Form Validation -------- */
+export function enableValidation(config) {
+  const formList = Array.from(document.querySelectorAll(config.formSelector));
+  formList.forEach((formElement) => {
+    setEventListeners(formElement, config);
+  });
+};
